requirement.txt
fastapi
uvicorn
pandas
gspread
google-auth
google-api-python-client
jinja2
python-multipart
